requirements.txt
aiogram==3.4.1
tortoise-orm==0.20.0